### 2021-03-02 #3044 @vissarion
:wave: @ysimillides, @HaoZeke, @yangbai90, @raghurama123, @Datseris, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
