### 2021-03-02 #3077 @whedon
[REVIEW]: HOODESolver.jl: A Julia package for highly oscillatory problems
**Submitting author:** @pnavaro (<a href="http://orcid.org/0000-0002-7372-3221">Pierre Navaro</a>)
**Repository:** <a href="https://github.com/ymocquar/HOODESolver.jl" target ="_blank">https://github.com/ymocquar/HOODESolver.jl</a>
**Version:** v0.2.3
