### 2021-04-19 ymocquar/HOODESolver.jl#7 @HaoZeke
Contribution Guidelines [JOSS]
Part of https://github.com/openjournals/joss-reviews/issues/3077.

I was wondering if the `README` could have a section added to it describing the contribution workflow (fork, open PR, ensure tests pass).  Some more explicit lines will go a long way in satisfying the `community guidelines` point. A few repo templates for issues would help as well.

However I consider this to be a non-blocker.
