### 2021-04-19 ymocquar/HOODESolver.jl#8 @HaoZeke
Minor Grammar Corrections [JOSS]
Part of https://github.com/openjournals/joss-reviews/issues/3077. In no order.
- [x] P1 L7 "has a long" --> "have a long|
- [x] P1 L11 "coming their" --> "relating to their"
- [x] P1 L12 "creating" --> "which gives rise to "
- [x] P1 L18 "method able" --> "method which is able"
- [x] P1 L23 "this type of stiff problems" --> "the stiff problems discussed above"
- [x] P1 L24 "solution" --> "solutions"
- [x] P1 L33: "uniform accuracy: uniformly accurate (UA) methods are indeed able to" --> "uniformly accurate (UA) methods which are"
- [x] P1 L33 "allows to skip several oscillations in a single time step" --> "allows skipping several oscillations each time step"
