### 2021-04-19 ymocquar/HOODESolver.jl#7 @pnavaro
Dear @ymocquar, I don't have access to repository settings to configure the issue template.
The documentation is available here
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#creating-issue-templates

I add an example here

---
name: Bug report
about: Create a report to help us improve
title: ' '
labels: bug
assignees: ' '
---

**Describe the bug**

A clear and concise description of what the bug is.

**Environment**

```julia
using HOODESolver
println(versioninfo())
using Pkg;
println(Pkg.status("HOODESolver"))
```

**Expected behavior**

A clear and concise description of what you expected to happen.
A minimal example of the input text should be supplied,
together with a screen-shot of the output document and/or command-line output, e.g.

**Additional context**

Add any other context about the problem here.

```

Cheers
