### 2021-04-19 ymocquar/HOODESolver.jl#8 @pnavaro
Thanks @HaoZeke for your comments, I updated the text in the paper branch.
