### 2021-04-19 ymocquar/HOODESolver.jl#8 @HaoZeke
Excellent, thanks @pnavaro
