### 2021-04-19 #3077 @HaoZeke
Hi @vissarion, I'm happy to recommend this for publication in JOSS in its current state. The package is well documented and well designed; it fills an important niche in the solution of ODE's by implementing the UA (uniformly accurate) methods. It is also developed with an eye towards being easy to use by users more used to existing Julia packages. Apologies for being so slow with this, and @pnavaro was very prompt to respond and handle all issues raised during the review.
