### 2021-04-20 ymocquar/HOODESolver.jl#9 @ranocha
Update dependency bounds
Thanks for this nice package! I wanted to install it but couldn't do that in my current setup because of version conflicts of dependencies of HOODESolver.jl and other direct dependencies of my project. It would be great if you could update the upper version bounds of

- GenericSchur.jl
- Polynomials.jl
- Reexport.jl

(as long as they are compatible with your package). You can automate this process using the GitHub action https://github.com/JuliaRegistries/CompatHelper.jl.
