### 2021-04-20 ymocquar/HOODESolver.jl#10 @ranocha
Examples from the docs do not work
Thanks for the nice documentation! However, some examples do not work properly, e.g. since `using Plots` is missing in

- https://ymocquar.github.io/HOODESolver.jl/stable/common_interface
- https://ymocquar.github.io/HOODESolver.jl/stable/quickstart

The latter yields additionally
```julia
julia> rng = MersenneTwister(1111)
ERROR: UndefVarError: MersenneTwister not defined
Stacktrace:
 [1] top-level scope
   @ REPL[19]:1
```

It would be great if you could turn these code examples into doctests using Documenter and check them automatically via GitHub actions.
