### 2021-04-20 ymocquar/HOODESolver.jl#11 @ranocha
JOSS paper
Thanks for creating this nice package, @pnavaro et al. It looks like a really valuable contribution to the community to me. Please find below some comments and suggestions.

- [x] Would it be possible to leverage ApproxFun.jl (and/or related Matlab tools) for these problems?
- [ ] From reading the manuscript and the README, it's not clear to me what kind of scalability to expect. Is it correct that this package is restricted to using a single CPU and only dense `Matrix` types for linear operator? At least something like the following (adapted from the README) does not work.
  ```julia
  julia> using HOODESolver, SparseArrays

  julia> A = spzeros(4, 4); A[1,3] = 1; A[3,1] = -1; A
  4×4 SparseMatrixCSC{Float64, Int64} with 2 stored entries:
    ⋅    ⋅   1.0   ⋅
    ⋅    ⋅    ⋅    ⋅
  -1.0   ⋅    ⋅    ⋅
    ⋅    ⋅    ⋅    ⋅

  julia> f1 = LinearHOODEOperator(1.0e-4, A)
  ERROR: MethodError: no method matching LinearHOODEOperator(::Float64, ::SparseMatrixCSC{Float64, Int64})
  Closest candidates are:
    LinearHOODEOperator(::T, ::Matrix{T} where T) where T at .../HOODESolver/src/common_interface.jl:9
  Stacktrace:
  [1] top-level scope
    @ REPL[3]:1
  ```
  Thus, it looks like the package is mostly made for (small) ODEs but not really applicable to, say, typical PDE discretizations that can run on a single workstation. It would be great if you could clarify this aspect.
