### 2021-04-20 ymocquar/HOODESolver.jl#10 @pnavaro
Thanks you @ranocha for your comments, I already use Documenter but some command are hidden though. I changed that.
