### 2021-04-20 ymocquar/HOODESolver.jl#9 @pnavaro
Thank you for noticing this oversight. I set up the GitHub action CompatHelper
