### 2021-04-20 ymocquar/HOODESolver.jl#11 @pnavaro
Thanks Hendrik,
- @crouseilles could you answer to the first question ?
- @ymocquar, could you add the method for this SparseMatrix case ?  I think we just need to replace Matrix{T} by AbstractMatrix{T}, but i am not sure there is no side effects.
