### 2021-04-20 ymocquar/HOODESolver.jl#11 @crouseilles
Salut Pierre,
je ne comprends pas trop les questions...
Nicolas.

> De: "Pierre Navaro" ***@***.***>
> À: "ymocquar" ***@***.***>
> Cc: "Nicolas Crouseilles" ***@***.***>, "Mention"
> ***@***.***>
> Envoyé: Mardi 20 Avril 2021 13:03:03
> Objet: Re: [ymocquar/HOODESolver.jl] JOSS paper (#11)

> Thanks Hendrik,

>    * [ https://github.com/crouseilles | @crouseilles ] could you answer to the
>     first question ?
>    * [ https://github.com/ymocquar | @ymocquar ] , could you add the method for
>    this SparseMatrix case ? I think we just need to replace Matrix{T} by
>     AbstractMatrix{T}, but i am not sure there si no side effects.
