### 2021-04-20 ymocquar/HOODESolver.jl#9 @ranocha
Looks like this is resolved. Thanks!
