### 2021-04-21 ymocquar/HOODESolver.jl#11 @pnavaro
- Answer to question 1 about the use of ApproxFun.jl

The numerical methods that are available in ApproxFun are undoubtedly efficient but they are not dedicated to solve solve stiff problems
with a uniformly accuracy. Indeed, for a fixed n (where n is the number of terms in the truncated series), one can find eps
such that the spectral accuracy of ApproxFun will be lost since a minimum number of points is required to sample one period).
On the contrary, our approach is able to capture the solution for arbitrary eps
with a uniform accuracy O(dt^p) with dt the time mesh and p the order of the method.
