### 2021-04-21 ymocquar/HOODESolver.jl#11 @ranocha
@pnavaro Sounds definitely interesting (and I se why that's the case and why you developed your methods). I guess other people not that familiar with the approach you use might have similar thoughts - maybe it would be worth pointing this out a bit more explicitly in your paper to stress the importance of your contribution?
