### 2021-04-21 ymocquar/HOODESolver.jl#11 @ranocha
In addition, ApproxFun uses an adaptive procedure to sample the solution while trying to deduce the optimal number of points. Thus, the oscillations (`epsilon` in your case) will make ApproxFun use more points. From your answer, I understand that you approach does not need so many points, correct?
Somehow related: You also use a spectral expansion in one pseudo-time variable `tau` [as described in the docs](https://ymocquar.github.io/HOODESolver.jl/stable/numerical_method/#Discretization). Would it be possible to use AppoxFun for that? Could there be any advantages or disadvantages?
