### 2021-04-26 ymocquar/HOODESolver.jl#10 @pnavaro
Changes are made in dev version of the documentation, it will be deploy to the stable version with the next release.

- https://ymocquar.github.io/HOODESolver.jl/dev/common_interface
- https://ymocquar.github.io/HOODESolver.jl/dev/quickstart
