### 2021-04-26 ymocquar/HOODESolver.jl#11 @crouseilles
> In addition, ApproxFun uses an adaptive procedure to sample the solution while trying to deduce the optimal number of points. Thus, the oscillations (`epsilon` in your case) will make ApproxFun use more points. From your answer, I understand that you approach does not need so many points, correct?

You are right, our approach dos not need to sample the oscillations with many points.

> Somehow related: You also use a spectral expansion in one pseudo-time variable `tau` [as described in the docs](https://ymocquar.github.io/HOODESolver.jl/stable/numerical_method/#Discretization). Would it be possible to use AppoxFun for that? Could there be any advantages or disadvantages?

In our context, the discretization of the periodic variable 'tau' is performed using fast Fourier transform.
If we understand correctly, ApproxFun is also doing Fourier expansion in the periodic framework.
Then, we think that the use of FFT is simpler than using ApproxFun which will add some dependencies.
