### 2021-04-26 ymocquar/HOODESolver.jl#11 @ranocha
Thanks for this clarification!
