### 2021-04-26 ymocquar/HOODESolver.jl#10 @ranocha
Thanks! Feel free to close this once all of the issues are resolved.
