### 2021-04-29 ymocquar/HOODESolver.jl#11 @pnavaro
>   Thus, it looks like the package is mostly made for (small) ODEs but not really applicable to, say, typical PDE discretizations that can run on a single workstation. It would be great if you could clarify this aspect.

I am not sure to understand your remark. In numerical method for plasma physics, you could have to follow particle trajectories by solving this kind of ODE. Our plan is to use this solver with

to handle a lot of particles with no need to use a small time step. We will try to use
[parallel ensemble feature](https://diffeq.sciml.ai/stable/features/ensemble/) of DifferentialEquations to do it. This is an ongoing work.
