### 2021-04-29 ymocquar/HOODESolver.jl#11 @ranocha
I was thinking of PDE discretizations that have so many degrees of freedom that a dense matrix of that size does not fit into the memory of a single workstation but sparse matrices can still be used to solve the problem on a workstation. Your kind of application focusing on charged particles in plasma physics leads to (o lot of) small ODE problems where you can use (small) dense matrices efficiently.
