### 2021-04-30 ymocquar/HOODESolver.jl#11 @pnavaro
I understand now. @ymocquar has implemented the SparseMatrix case. I will trigger the release. We tried to increase the number of variables and we did not see any issues. I think it is possible to solve Maxwell equation using finite element discretization and get the suitable equation for time evolution that can be solved with HOODE. There is some very specific condition on matrix A. It needs some time to write the test case though.

PS: I dont receive the notification when your answer, i have to investigate that so sorry for late answer.
