### 2021-04-30 ymocquar/HOODESolver.jl#11 @ranocha
Don't worry, I'm not in a rush. Thanks for the clarification. Feel free to close this issue once everything is ready and tests using sparse matrices are running in a new release.
