### 2021-05-01 ymocquar/HOODESolver.jl#11 @pnavaro
fix in v0.2.4
