### 2021-05-07 #3077 @pnavaro
> Thanks for creating this nice package, @pnavaro et al. It looks like a really valuable contribution to the community to me.
>
> * [ ]  [ymocquar/HOODESolver.jl#9](https://github.com/ymocquar/HOODESolver.jl/issues/9)
> * [ ]  [ymocquar/HOODESolver.jl#10](https://github.com/ymocquar/HOODESolver.jl/issues/10)
> * [ ]  [ymocquar/HOODESolver.jl#11](https://github.com/ymocquar/HOODESolver.jl/issues/11)
>
> Nevertheless, this contribution looks already really good and I will be happy to recommend publication once the issues described above are resolved.

All issues are fixed. Thanks for your comments @ranocha
