### 2021-05-10 #3077 @vissarion
> > Yes, please change the title and the list of authors in zenodo to match the ones in your paper.
>
> Sorry, it's done.

great, please also change the title to "HOODESolver.jl: A Julia package for highly oscillatory problems" instead of "ymocquar/HOODESolver.jl: v0.2.4"
