### 2021-05-10 #3077 @pnavaro
> @pnavaro great, please also change the title to "HOODESolver.jl: A Julia package for highly oscillatory problems" instead of "ymocquar/HOODESolver.jl: v0.2.4"

Since i am not the repository owner, it takes some time to make the changes but we get there.
