### 2021-05-14 #3077 @danielskatz
@vissarion - the command you need here is `@whedon accept from branch paper`

I don't want to do this for you, since I want you to be able to check the results and let `@openjournals/joss-eics` know when this is ready to proceed
