### 2021-05-14 #3077 @vissarion
@pnavaro I proposed some very minor editing suggestions https://github.com/ymocquar/HOODESolver.jl/pull/16
@danielskatz after addressing the above issue it is ready from my side
