### 2021-05-17 #3077 @pnavaro
> @pnavaro I proposed some very minor editing suggestions [ymocquar/HOODESolver.jl#16](https://github.com/ymocquar/HOODESolver.jl/pull/16)
> @danielskatz after addressing the above issue it is ready from my side

Ok merged. Thanks @vissarion
