### 2021-05-18 #3077 @kyleniemeyer
@vissarion moving forward, you can indicate that the submission is ready to accept using `@whedon accept`, which does some work and also notifies the EIC team
