### 2021-05-18 #3077 @vissarion
> @vissarion moving forward, you can indicate that the submission is ready to accept using `@whedon accept`, which does some work and also notifies the EIC team

Sure,  I have done it above but I should have done it again after the changes in the paper. Thanks for the tip!
