### 2021-05-18 #3077 @kyleniemeyer
@pnavaro Hi there, I'm doing some final checks before accepting your submission. Just to confirm, is the paper still in the `paper` branch, or on the main branch now?
