### 2021-05-18 #3077 @kyleniemeyer
@pnavaro you did not have to merge that in, I just wanted to make sure I was using the correct source for publishing the paper. You can remove the paper from your `master` branch, if you'd like, and I will use the `paper` branch.
