### 2021-05-18 #3077 @kyleniemeyer
Hi @pnavaro, sorry for a final change, but I just noticed that the Chartier et al. reference ("Derivative-free high-order...") is actually in press at a journal, and should include that information instead of being referenced as an arXiv eprint. Can you add the journal and update the year to 2021?
